Too much ROBA
This is the last Italian thing in the project because... I translated every comment to english yayyyy!
Anyway i made a fix to the timer system and many little changes, but the big thing of this comment is the language, finally the project is all in english.
(Except for this commit title: roba mean "stuff" in Italian :P)